Send cost as part of signed payload
[sc-76277]
