+++ brownie-config.yaml
@@ -5,10 +5,12 @@ dependencies:
 compiler:
   solc:
     remappings:
-      - '@openzeppelin=OpenZeppelin/openzeppelin-contracts@4.5.0'
-      - '@openzeppelin-upgradeable=OpenZeppelin/openzeppelin-contracts-upgradeable@4.5.0'
+      - "@openzeppelin=OpenZeppelin/openzeppelin-contracts@4.5.0"
+      - "@openzeppelin-upgradeable=OpenZeppelin/openzeppelin-contracts-upgradeable@4.5.0"
 networks:
   ropsten:
+    verify: True
+  rinkeby:
     verify: True
   mainnet-fork:
     verify: False
